Revert "[TDF] Avoid copies when reading values through a data-source"
This reverts commit b781f187ec9b49e64f939e2b76bf993d9356ed41.

We needed to do this as this commit introduces also a bug. Somehow the
(names of) the column types are not correctly taken into account.
As a result errors like this one were printed:
'''
terminate called after throwing an instance of 'std::runtime_error'
  what():  TColumnValue: type specified for column "col0" is y but temporary column has type Py
Aborted (core dumped)
'''
Given that this is a performance optimisation, despite being extremely important,
I tend to disable it for the moment and to come back to it once the TDS functionalities
are all added to the codebase.